refactor: Refactor TabsContainer into files
